finall install mysql and connect into my application

+++ service-app/src/entity/product.entity.ts
@@ -13,6 +13,6 @@ export class Product {
   })
   description: string;
 
-  @Column('decimal')
+  @Column('float')
   price: number;
 }
